examples/generate_pdf/requirements.txt
pycairo
nicegui